qa/deepsea: separate ses_qa.py task for non-DS testing
(manual cherry-pick of 62713a25e1ff8896f48ef68b4cfa74a4d2098101)

Conflicts:
- do not introduce deepsea/tier2/mgr-plugins subsuite
